+++ src/main/java/racingcar/controller/RacingController.java
@@ -1,5 +1,6 @@
 package racingcar.controller;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -10,6 +11,7 @@ import racingcar.domain.Car;
 import racingcar.domain.Cars;
 import racingcar.domain.GameCount;
 import racingcar.domain.PowerGenerator;
+import racingcar.service.RacingCarService;
 import racingcar.util.CarNamesDivider;
 
 import java.util.List;
@@ -20,36 +22,17 @@ import static java.util.stream.Collectors.*;
 @RestController
 public class RacingController {
 
-    private final CarNamesDivider carNamesDivider;
+    private final RacingCarService racingCarService;
 
-    public RacingController() {
-        carNamesDivider = new CarNamesDivider();
+    @Autowired
+    public RacingController(RacingCarService racingCarService) {
+        this.racingCarService = racingCarService;
     }
 
     @PostMapping("/plays")
     public ResponseEntity<PlaysResponseDto> run(@RequestBody PlaysRequestDto playsRequestDto) {
-        List<String> carNamesByDivider = carNamesDivider.divideCarNames(playsRequestDto.getNames());
-        List<Car> inputCars = carNamesByDivider.stream()
-                .map(Car::new)
-                .collect(toList());
-        Cars cars = new Cars(inputCars);
-        GameCount gameCount = new GameCount(playsRequestDto.getCount());
-        progress(cars, gameCount);
-        String winners = cars.getWinners().stream()
-                .map(Car::getName)
-                .collect(joining(","));
-        return ResponseEntity.ok().body(new PlaysResponseDto(winners, cars.getCars()));
-    }
-
-    private void progress(Cars cars, GameCount gameCount) {
-        while (gameCount.isGameProgress()) {
-            gameCount.proceedOnce();
-            moveAllCar(cars);
-        }
-    }
-
-    private void moveAllCar(Cars cars) {
-        cars.moveAll(new PowerGenerator(new Random()));
+        return ResponseEntity.ok()
+                .body(racingCarService.plays(playsRequestDto));
     }
 
 }
